refactor: move conversational utils from cli to eval module
Move LegacyEvalChatMessagesMapper and related models from
uipath._cli._evals._conversational_utils to
uipath.eval.models._conversational_utils so the eval module
has no dependency on the cli namespace.
