+++ src/uipath/eval/models/evaluation_set.py
@@ -5,15 +5,14 @@ from typing import Any, Literal
 from pydantic import BaseModel, ConfigDict, Field
 from pydantic.alias_generators import to_camel
 
-from uipath._cli._evals._conversational_utils import (
-    LegacyConversationalEvalInput,
-    LegacyConversationalEvalOutput,
-)
-
 from ..mocks._types import (
     InputMockingStrategy,
     MockingStrategy,
     ToolSimulation,
+)
+from ._conversational_utils import (
+    LegacyConversationalEvalInput,
+    LegacyConversationalEvalOutput,
 )
