+++ tests/cli/eval/test_conversational_utils.py
@@ -1,6 +1,7 @@
 """Tests for conversational eval utilities."""
 
-from uipath._cli._evals._conversational_utils import (
+from uipath.core.chat import UiPathInlineValue
+from uipath.eval.models._conversational_utils import (
     LegacyConversationalEvalInput,
     LegacyConversationalEvalInputAgentMessage,
     LegacyConversationalEvalInputToolCall,
@@ -11,7 +12,6 @@ from uipath._cli._evals._conversational_utils import (
     LegacyConversationalEvalUserMessage,
     UiPathLegacyEvalChatMessagesMapper,
 )
-from uipath.core.chat import UiPathInlineValue
 
 
 class TestLegacyConversationalEvalInputToUiPathMessages:
